requirements.txt
discord
python-dotenv
requests
giphy_client
